Time: 9 ms (21.31%), Space: 13.7 MB (45.41%) - LeetHub

+++ 0330-patching-array/0330-patching-array.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int minPatches(vector<int>& nums, int n) {
-        int maxReach = 0;
+        long maxReach = 0;
         int patch = 0;
         int i = 0;
         while (maxReach < n) {
